requirements.txt
pymupdf>=1.24.0
sentence-transformers>=5.0.0
chromadb==1.5.0
rank-bm25>=0.2.2
nltk>=3.8.0
groq>=1.0.0
streamlit>=1.45.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
numpy>=1.26.0
faster-whisper>=1.0.0
soundfile>=0.12.1
transformers>=4.51.0
torch>=2.0.0
huggingface_hub>=0.23.0
contextual-client>=0.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
requests>=2.31.0
qdrant-client>=1.11.1
